separated observer from agent for use in planned replay lobe
