+++ examples/examples-common/src/terran_lobe.rs
@@ -151,7 +151,7 @@ impl InGame {
     fn update(mut self, msg: Protocol<Message, Role>) -> Result<TerranLobe> {
         if let Some(msg) = self.soma.update(msg)? {
             match msg {
-                Protocol::Message(_, Message::Update(frame)) => {
+                Protocol::Message(_, Message::Observation(frame)) => {
                     self.on_frame(frame)
                 },
